add authentication utils
update linter
add login page redirection
private page router reworked
