+++ .eslintrc.json
@@ -1,6 +1,8 @@
 {
   "extends": "react-app",
   "rules": {
+    "padded-blocks": ["error", "never"],
+    "prefer-const": "error",
     "padding-line-between-statements": [
       "error",
       { "blankLine": "always", "prev": ["const", "let", "var"], "next": "*"},
@@ -44,6 +46,12 @@
     "semi": [
       "error",
       "always"
-    ]
+    ],
+    "react/jsx-tag-spacing": ["error", {
+      "closingSlash": "never",
+      "beforeSelfClosing": "always",
+      "afterOpening": "never",
+      "beforeClosing": "never"
+    }]
   }
 }
