+++ src/components/index.js
@@ -2,3 +2,5 @@ export { Label } from './label.component';
 export { Subtitle } from './subtitle.component';
 export { Title } from './title.component';
 export { Header } from './header.components';
+export { Input } from './input.component';
+export { Button } from './button.component';
