+++ src/pages/register-page/components/register-form.components.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Redirect } from 'react-router-dom';
 import { Form, Field } from 'react-final-form';
 import { register } from '../../../services/authentication.services';
-import { LocalStorage } from '../../../utils/storages.utils';
+import { Authentication } from '../../../utils';
 
 const onSubmit = async (values) => {
     await register(values);
@@ -13,10 +13,7 @@ const validate = () => {
 };
 
 export const RegisterForm = () => {
-
-    const token = LocalStorage.getItem('auth_token');
-
-    if (token != null) {
+    if (!Authentication.isAuthenticated()) {
         return (
             <Redirect to="/" />
         );
@@ -28,17 +25,14 @@ export const RegisterForm = () => {
             validate={ validate }
             render={({ handleSubmit }) => (
                 <form onSubmit={ handleSubmit }>
-
                     <div>
                         <label>Username</label>
                         <Field name="username" component="input" type="text" placeholder="username" />
                     </div>
-
                     <div>
                         <label>Password</label>
                         <Field name="password" component="input" type="password" placeholder="password" />
                     </div>
-
                     <div>
                         <label>Role</label>
                         <Field name="role" component="select">
@@ -48,9 +42,7 @@ export const RegisterForm = () => {
                             <option value="TRADER">Trader</option>
                         </Field>
                     </div>
-
                     <button type="submit">Submit</button>
-
                 </form>
             )}
         />
