+++ src/utils/index.js
@@ -1,3 +1,4 @@
 export { Errors } from './errors.utils';
 export { simpleFetch } from './fetch.utils';
 export { LocalStorage, SessionStorage } from './storages.utils';
+export { Authentication } from './authentication.utils';
